[update] fixed bugs

+++ internal/interface/repository/user.go
@@ -4,7 +4,7 @@ import "clean_architecture_with_ddd/internal/entity"
 
 type UserRepository interface {
 	GetUserByUsername(username string) (*entity.User, error)
-	GetTaxFromUserByTaxId(userId int) (float32, error)
+	GetTaxFromUserByTaxId(userId int) (int, error)
 
 	CreateUser(user *entity.User) error
 }
